requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cssselect==1.1.0
dateutils==0.6.12
demjson==2.2.4
Django==4.0.4
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
lxml==4.9.1
MarkupSafe==2.1.1
pyasn1==0.4.8
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.27.1
rsa==4.8
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.7.0
